Re-organize demo app packages
