+++ app/src/main/java/com/kenkeremath/uselessui/app/ui/components/ParameterSlider.kt
@@ -1,0 +1,50 @@
+import android.annotation.SuppressLint
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.Slider
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import java.math.RoundingMode
+import java.text.DecimalFormat
+
+@SuppressLint("DefaultLocale")
+@Composable
+internal fun ParameterSlider(
+    label: String,
+    value: Float,
+    onValueChange: (Float) -> Unit,
+    valueRange: ClosedFloatingPointRange<Float>,
+    modifier: Modifier = Modifier,
+    showAsInt: Boolean = false,
+) {
+
+    val intFormat = remember {
+        val df = DecimalFormat("#")
+        df.roundingMode = RoundingMode.CEILING
+        df
+    }
+    Column(modifier = modifier.fillMaxWidth()) {
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            horizontalArrangement = Arrangement.SpaceBetween
+        ) {
+            Text(text = label)
+            val formattedValue = if (!showAsInt) {
+                String.format("%.1f", value)
+            } else {
+                intFormat.format(value)
+            }
+            Text(text = formattedValue)
+        }
+        Slider(
+            value = value,
+            onValueChange = onValueChange,
+            valueRange = valueRange,
+            modifier = Modifier.fillMaxWidth()
+        )
+    }
+}
